Slett gammel kode
