--- src/bedriftsmeny/Virksomhetsvelger/Menyvalg/Sokefelt/Sokefelt.css
@@ -1,8 +1,0 @@
-.virksomhetsvelger .bedriftsmeny-sokefelt {
-    margin: 1rem;
-}
-
-.virksomhetsvelger .bedriftsmeny-sokefelt__skjult-aria-live-sokeresultat {
-    font-style: italic;
-    margin: 1rem 0 0 1rem;
-}
