--- src/bedriftsmeny/Virksomhetsvelger/Menyvalg/Underenhetsvelger/Organisasjonsbeskrivelse/Organisasjonsbeskrivelse.css
@@ -1,38 +1,0 @@
-.organisasjonsbeskrivelse {
-    display: flex;
-    align-items: center;
-    width: 16rem;
-}
-
-@media (min-width: 375px) {
-    .organisasjonsbeskrivelse {
-        width: 18.5rem;
-    }
-}
-
-@media (min-width: 48rem) {
-    .organisasjonsbeskrivelse {
-        width: 22rem;
-    }
-}
-
-.organisasjonsbeskrivelse__ikon {
-    height: 2rem;
-    width: 2rem;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin-right: 0.5rem;
-    flex-shrink: 0;
-}
-
-.organisasjonsbeskrivelse__beskrivelse {
-    overflow: hidden;
-    text-align: left;
-}
-
-.organisasjonsbeskrivelse__navn {
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    overflow: hidden;
-}
