--- src/bedriftsmeny/Virksomhetsvelger/Menyvalg/Underenhetsvelger/Underenhet/Underenhet.css
@@ -1,43 +1,0 @@
-.underenhet {
-    width: 100%;
-    border-top: 1px solid #979797;
-    cursor: pointer;
-}
-
-.underenhet.valgtunderenhet {
-    background-color: var(--navds-semantic-color-interaction-primary);
-    color: white;
-}
-
-.underenhet.valgtunderenhet svg g {
-    fill: none;
-    stroke: white;
-}
-
-.underenhet.valgtunderenhet-grey-on-hover {
-    background-color: var(--navds-global-color-gray-200);
-    color: var(--navds-semantic-color-text);
-}
-
-.underenhet.valgtunderenhet-grey-on-hover svg g {
-    fill: var(--navds-semantic-color-text);
-    stroke: white;
-}
-
-.underenhet:focus {
-    box-shadow: inset 0 0 0 3px #254b6d;
-    outline: none;
-}
-
-.underenhet .organisasjonsbeskrivelse__ikon {
-    margin-left: 0.5rem;
-}
-
-.underenhet:hover {
-    color: white;
-    background-color: var(--navds-semantic-color-interaction-primary);
-}
-
-.underenhet:hover svg g {
-    stroke: white;
-}
