+++ src/bedriftsmeny/Virksomhetsvelger/Virksomhetsvelger.tsx
@@ -1,170 +1,116 @@
 import React, { FunctionComponent, useContext, useEffect, useRef, useState } from 'react';
-
-import {
-    JuridiskEnhetMedUnderEnheterArray,
-    Organisasjon,
-    tomAltinnOrganisasjon,
-} from '../organisasjon';
-import Menyvalg from './Menyvalg/Menyvalg';
-import Sokefelt from './Menyvalg/Sokefelt/Sokefelt';
-import MenyKnapp from './Menyknapp/Menyknapp';
-import { setfokusPaMenyKnapp, setfokusPaSokefelt } from './Menyvalg/pilnavigerinsfunksjoner';
-import { useHandleOutsideEvent } from './useHandleOutsideEvent';
+import { Button, Popover, Heading, BodyShort, Search, Accordion, Detail } from '@navikt/ds-react';
+import { Organisasjon } from '../organisasjon';
+import { Expand, Collapse, Office1 } from '@navikt/ds-icons';
 import { VirksomhetsvelgerContext } from './VirksomhetsvelgerProvider';
+import JuridiskEnhet from './JuridiskEnhet';
 
-export interface VirksomhetsvelgerProps {
+type Props = {
     onOrganisasjonChange?: (organisasjon: Organisasjon) => void;
-}
-
-const finnForsteJuridiskEnhet = (
-    organisasjonstre: JuridiskEnhetMedUnderEnheterArray[],
-    soketekst: string,
-    orgNrValgtJuridiskEnhet: string
-): Organisasjon => {
-    if (organisasjonstre.length === 0) {
-        return tomAltinnOrganisasjon;
-    }
-
-    if (soketekst.length > 0) {
-        return organisasjonstre[0].JuridiskEnhet;
-    }
-
-    const juridiskeEnheter = organisasjonstre.map(({ JuridiskEnhet }) => JuridiskEnhet);
-    return (
-        juridiskeEnheter.find(
-            ({ OrganizationNumber }) => OrganizationNumber === orgNrValgtJuridiskEnhet
-        ) ?? tomAltinnOrganisasjon
-    );
 };
 
-const Virksomhetsvelger: FunctionComponent<VirksomhetsvelgerProps> = (props) => {
-    const bedriftvelgernode = useRef<HTMLDivElement>(null);
-    const { onOrganisasjonChange } = props;
+const Velger: FunctionComponent<Props> = ({ onOrganisasjonChange }) => {
+    const buttonRef = useRef<HTMLButtonElement>(null);
+    const searchRef = useRef<HTMLInputElement>(null);
+    const [åpen, setÅpen] = useState<boolean>(false);
+
     const {
+        velgUnderenhet,
         valgtOrganisasjon,
         aktivtOrganisasjonstre,
-        velgUnderenhet,
-        søketekst: soketekst,
+        søketekst,
+        setSøketekst,
     } = useContext(VirksomhetsvelgerContext);
-    const [erApen, setErApen] = useState(false);
 
-    const [organisasjonIFokus, setOrganisasjonIFokus] = useState(tomAltinnOrganisasjon);
-    const [forrigeOrganisasjonIFokus, setForrigeOrganisasjonIFokus] = useState(
-        tomAltinnOrganisasjon
+    const toggleVelger = (verdi?: boolean) => {
+        setÅpen(verdi === undefined ? !åpen : verdi);
+    };
+
+    const onUnderenhetClick = (virksomhet: Organisasjon) => {
+        if (onOrganisasjonChange) {
+            virksomhet;
+        }
+
+        velgUnderenhet(virksomhet.OrganizationNumber);
+        setÅpen(false);
+    };
+
+    useEffect(() => {
+        if (åpen) {
+            searchRef.current?.focus();
+        }
+    }, [åpen]);
+
+    const antallTreff = aktivtOrganisasjonstre.reduce(
+        (totaltAntall, juridiskEnhet) => totaltAntall + juridiskEnhet.Underenheter.length,
+        0
     );
 
-    useEffect(() => {
-        setErApen(false);
-        onOrganisasjonChange?.(valgtOrganisasjon);
-    }, [valgtOrganisasjon]);
-
-    useEffect(() => {
-        if (!erApen) {
-            setfokusPaMenyKnapp();
-            setOrganisasjonIFokus(tomAltinnOrganisasjon);
-        } else {
-            setfokusPaSokefelt();
-        }
-    }, [erApen]);
-
-    useHandleOutsideEvent(bedriftvelgernode, () => {
-        setErApen(false);
-    });
-
-    const onEnterSearchbox = () => {
-        if (soketekst.length > 0 && aktivtOrganisasjonstre.length > 0) {
-            const kunTreffPåEnUnderenhet =
-                aktivtOrganisasjonstre.length === 1 &&
-                aktivtOrganisasjonstre[0].Underenheter.length === 1;
-            if (kunTreffPåEnUnderenhet) {
-                const underenhet = aktivtOrganisasjonstre[0].Underenheter[0];
-                if (underenhet.OrganizationNumber !== valgtOrganisasjon.OrganizationNumber) {
-                    velgUnderenhet(aktivtOrganisasjonstre[0].Underenheter[0].OrganizationNumber);
-                } else {
-                    setErApen(false);
-                }
-            } else {
-                setOrganisasjonIFokus(aktivtOrganisasjonstre[0].JuridiskEnhet);
-            }
-        }
-    };
-
-    const setFocusOnForsteVirksomhet = () => {
-        if (aktivtOrganisasjonstre.length > 0 || soketekst.length === 0) {
-            const forsteJuridiskEnhetILista = finnForsteJuridiskEnhet(
-                aktivtOrganisasjonstre,
-                soketekst,
-                valgtOrganisasjon.ParentOrganizationNumber
-            );
-            const blarOppTilSøkefeltOgNedTilMeny =
-                forrigeOrganisasjonIFokus.OrganizationNumber ===
-                aktivtOrganisasjonstre[0].JuridiskEnhet.OrganizationNumber;
-            const valgtJuridiskEnhetErFørsteILista =
-                forsteJuridiskEnhetILista.OrganizationNumber ===
-                aktivtOrganisasjonstre[0].JuridiskEnhet.OrganizationNumber;
-            const skalBlaTilFørsteElementIMenyKomponenter =
-                (blarOppTilSøkefeltOgNedTilMeny && !valgtJuridiskEnhetErFørsteILista) ||
-                soketekst.length > 0;
-
-            if (skalBlaTilFørsteElementIMenyKomponenter) {
-                setOrganisasjonIFokus(aktivtOrganisasjonstre[0].JuridiskEnhet);
-            } else {
-                setOrganisasjonIFokus(forsteJuridiskEnhetILista);
-            }
-        }
-    };
-
-    const antallTreff = aktivtOrganisasjonstre
-        .map(({ Underenheter }) => Underenheter.length)
-        .reduce((x, y) => x + y, 0);
-
     return (
-        <nav
-            className="virksomhetsvelger"
-            aria-label="Velg virksomhet"
-            onKeyDown={(event) => {
-                if (event.key === 'Escape' || event.key === 'Esc') {
-                    setErApen(false);
-                }
-            }}
-        >
-            <div ref={bedriftvelgernode}>
-                <MenyKnapp erApen={erApen} setErApen={setErApen} />
-                <div
-                    role="toolbar"
-                    className={`virksomhetsvelger__dropdown--${erApen ? 'apen' : 'lukket'}`}
-                    aria-hidden={!erApen}
-                    id="virksomhetsvelger__dropdown"
-                >
-                    <Sokefelt
-                        onArrowUp={() => setfokusPaMenyKnapp()}
-                        onArrowDown={() => setFocusOnForsteVirksomhet()}
-                        onEnter={() => onEnterSearchbox()}
-                        antallTreff={antallTreff}
-                    />
-
-                    <div className="dropdownmeny-elementer-wrapper">
-                        <div
-                            className={`dropdownmeny-elementer ${
-                                !!soketekst ? 'medSokeTekst' : ''
-                            }`}
-                        >
-                            {aktivtOrganisasjonstre.length > 0 && (
-                                <Menyvalg
-                                    organisasjonIFokus={organisasjonIFokus}
-                                    setOrganisasjonIFokus={setOrganisasjonIFokus}
-                                    forrigeOrganisasjonIFokus={forrigeOrganisasjonIFokus}
-                                    setForrigeOrganisasjonIFokus={setForrigeOrganisasjonIFokus}
-                                    erApen={erApen}
-                                    setErApen={setErApen}
-                                />
-                            )}
-                        </div>
+        <>
+            <Button
+                className="navbm-virksomhetsvelger"
+                onClick={() => toggleVelger()}
+                type="button"
+                variant="secondary"
+                ref={buttonRef}
+                role="menubutton"
+                aria-label="Velg aktiv virksomhet"
+                aria-controls="navbm-virksomhetsvelger-popup"
+                aria-haspopup={true}
+                aria-expanded={åpen}
+            >
+                <div className="navbm-virksomhetsvelger__innhold">
+                    <Office1 aria-hidden={true} />
+                    <div className="navbm-virksomhetsvelger__tekst">
+                        <Heading size="small" level="2">
+                            {valgtOrganisasjon.Name}
+                        </Heading>
+                        <BodyShort>virksomhetsnr. {valgtOrganisasjon.OrganizationNumber}</BodyShort>
                     </div>
+                    {åpen ? <Collapse aria-hidden={true} /> : <Expand aria-hidden={true} />}
                 </div>
-            </div>
-        </nav>
+            </Button>
+            <Popover
+                offset={24}
+                open={åpen}
+                onClose={() => setÅpen(false)}
+                anchorEl={buttonRef.current}
+                placement="bottom-start"
+                id="navbm-virksomhetsvelger-popup"
+            >
+                <div className="navbm-virksomhetsvelger__popup" role="menu">
+                    <Search
+                        ref={searchRef}
+                        variant="simple"
+                        value={søketekst}
+                        onChange={setSøketekst}
+                        placeholder="Søk på virksomhet ..."
+                        label="Søk på virksomhet"
+                    />
+                    {søketekst.length > 0 && (
+                        <Detail aria-live="polite">
+                            {antallTreff === 0 ? 'Ingen' : antallTreff} treff på "{søketekst}"
+                        </Detail>
+                    )}
+                    <Accordion>
+                        <ul
+                            role="presentation"
+                            className="navbm-virksomhetsvelger__juridiske-enheter"
+                        >
+                            {aktivtOrganisasjonstre.map((juridiskEnhet) => (
+                                <JuridiskEnhet
+                                    juridiskEnhet={juridiskEnhet}
+                                    valgtOrganisasjon={valgtOrganisasjon}
+                                    onUnderenhetClick={onUnderenhetClick}
+                                />
+                            ))}
+                        </ul>
+                    </Accordion>
+                </div>
+            </Popover>
+        </>
     );
 };
 
-export default Virksomhetsvelger;
+export default Velger;
