+++ src/bedriftsmeny/Virksomhetsvelger/utils.ts
@@ -1,5 +1,5 @@
-import {Organisasjon, JuridiskEnhetMedUnderEnheterArray} from '../../organisasjon';
-import {useLocation, useNavigate} from "react-router-dom";
+import { Organisasjon, JuridiskEnhetMedUnderEnheterArray } from '../organisasjon';
+import { useLocation, useNavigate } from 'react-router-dom';
 
 const ORGNUMMER_PARAMETER = 'bedrift';
 const ORGNUMMER_LOCAL_STORE = 'virksomhetsvelger_bedrift';
@@ -14,17 +14,17 @@ export const hentUnderenheter = (organisasjonstre: JuridiskEnhetMedUnderEnheterA
     organisasjonstre.reduce(
         (organisasjoner: Organisasjon[], parentOrg) => [
             ...organisasjoner,
-            ...parentOrg.Underenheter
+            ...parentOrg.Underenheter,
         ],
         []
     );
 
-export type OrgnrSearchParamType = () => [string | null, (orgnr: string) => void]
+export type OrgnrSearchParamType = () => [string | null, (orgnr: string) => void];
 
 export const useOrgnrSearchParam: OrgnrSearchParamType = () => {
     const location = useLocation();
-    const searchParam = new URLSearchParams(location.search)
-    const currentOrgnr = searchParam.get(ORGNUMMER_PARAMETER)
+    const searchParam = new URLSearchParams(location.search);
+    const currentOrgnr = searchParam.get(ORGNUMMER_PARAMETER);
     const navigate = useNavigate();
 
     const setOrgnr = (orgnr: string) => {
@@ -34,9 +34,9 @@ export const useOrgnrSearchParam: OrgnrSearchParamType = () => {
             } else {
                 searchParam.set(ORGNUMMER_PARAMETER, orgnr);
             }
-            navigate({search: searchParam.toString()})
+            navigate({ search: searchParam.toString() });
         }
-    }
+    };
 
-    return [currentOrgnr, setOrgnr]
-}
+    return [currentOrgnr, setOrgnr];
+};
