+++ src/bedriftsmeny/index.css
@@ -1,10 +1,3 @@
 @import './BedriftsmenyView.css';
-@import './Virksomhetsvelger/Virksomhetsvelger.css';
-@import './Virksomhetsvelger/Menyknapp/MenyknappView.css';
-@import './Virksomhetsvelger/Menyvalg/Sokefelt/Sokefelt.css';
-@import './Virksomhetsvelger/Menyvalg/Underenhetsvelger/Underenhetsvelger.css';
-@import './Virksomhetsvelger/Menyvalg/Underenhetsvelger/Organisasjonsbeskrivelse/Organisasjonsbeskrivelse.css';
-@import './Virksomhetsvelger/Menyvalg/Underenhetsvelger/Underenhet/Underenhet.css';
-
-@import './velger/Virksomhetsvelger.css';
-@import './velger/JuridiskEnhet.css';
+@import './virksomhetsvelger/Virksomhetsvelger.css';
+@import './virksomhetsvelger/JuridiskEnhet.css';
