--- src/bedriftsmeny/velger/Virksomhetsvelger.css
@@ -1,53 +1,0 @@
-.navbm-virksomhetsvelger + .navds-popover .navds-popover__arrow {
-    background-color: var(--navds-semantic-color-canvas-background);
-}
-
-.navbm-virksomhetsvelger:active {
-    color: white;
-}
-
-.navbm-virksomhetsvelger__innhold {
-    display: grid;
-    grid-template-columns: auto 1fr auto;
-    align-items: center;
-    gap: 1rem;
-}
-
-.navbm-virksomhetsvelger__innhold svg {
-    width: 1.25rem;
-    height: 1.25rem;
-}
-
-.navbm-virksomhetsvelger__tekst {
-    display: flex;
-    flex-direction: column;
-    text-align: left;
-    width: 15rem;
-}
-
-.navbm-virksomhetsvelger__tekst h2 {
-    white-space: pre;
-    text-overflow: ellipsis;
-    overflow: hidden;
-}
-
-.navbm-virksomhetsvelger__popup {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
-    width: 26.25rem;
-    padding: var(--navds-spacing-3) var(--navds-spacing-3);
-    background-color: var(--navds-semantic-color-canvas-background);
-    border-radius: var(--navds-border-radius-small);
-}
-
-.navbm-virksomhetsvelger__juridiske-enheter {
-    padding: 0;
-    margin: 0;
-
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-    max-height: 65vh;
-    overflow-y: scroll;
-}
